requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# OpenAI
openai>=1.12.0

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.26.3

# YAML Processing
pyyaml>=6.0.1

# HTTP Client
httpx>=0.26.0

# Logging
python-json-logger>=2.0.7

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Type checking
mypy>=1.8.0
